# dashboard/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchRecord": {
        "description": "\u6279\u91cf\u5b9e\u9a8c\u8bb0\u5f55\uff0c\u4fdd\u5b58 manifest\u3001\u8fdb\u5ea6\u3001run IDs \u548c\u5b89\u5168\u58f0\u660e\u3002",
        "properties": {
          "artifact_paths": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Artifact Paths",
            "type": "object"
          },
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "hardware_write_operations": {
            "items": {
              "type": "string"
            },
            "title": "Hardware Write Operations",
            "type": "array"
          },
          "manifest": {
            "$ref": "#/components/schemas/ExperimentManifest"
          },
          "progress": {
            "$ref": "#/components/schemas/BatchProgress"
          },
          "run_ids": {
            "items": {
              "type": "string"
            },
            "title": "Run Ids",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/SimulationRunStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": ["batch_id", "manifest", "progress", "run_ids", "status"],
        "title": "BatchRecord",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "BIG-small Cloud Planning API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/simulation/batches": {
      "post": {
        "operationId": "create_batch_api_v1_simulation_batches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Batch",
        "tags": ["simulation-workbench"]
      }
    },
    "/api/v1/simulation/batches/{batch_id}": {
      "get": {
        "operationId": "get_batch_api_v1_simulation_batches__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Batch",
        "tags": ["simulation-workbench"]
      }
    },
    "/api/v1/simulation/batches/{batch_id}/cancel": {
      "post": {
        "operationId": "cancel_batch_api_v1_simulation_batches__batch_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Batch",
        "tags": ["simulation-workbench"]
      }
    },
    "/api/v1/simulation/batches/{batch_id}/retry-failed": {
      "post": {
        "operationId": "retry_failed_batch_api_v1_simulation_batches__batch_id__retry_failed_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Failed Batch",
        "tags": ["simulation-workbench"]
      }
    },
    "/api/v1/simulation/batches/{batch_id}/runs": {
      "get": {
        "operationId": "batch_runs_api_v1_simulation_batches__batch_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Runs",
        "tags": ["simulation-workbench"]
      }
    },
...
    "/api/v1/simulation/comparisons": {
      "post": {
        "operationId": "comparisons_api_v1_simulation_comparisons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparisonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cloud_edge_robot_arm__simulation_workbench__models__ComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Comparisons",
        "tags": ["simulation-workbench"]
      }
    },
    "/api/v1/simulation/exports": {
      "post": {
        "operationId": "exports_api_v1_simulation_exports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exports",
        "tags": ["simulation-workbench"]
      }
    },
...
    "/api/v1/simulation/runs": {
      "get": {
        "operationId": "list_runs_api_v1_simulation_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Runs",
        "tags": ["simulation-workbench"]
      },
      "post": {
        "operationId": "create_run_api_v1_simulation_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": ["simulation-workbench"]
      }
    },
    "/api/v1/simulation/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_simulation_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": ["simulation-workbench"]
      }
    },
    "/api/v1/simulation/runs/{run_id}/artifacts": {
      "get": {
        "operationId": "run_artifacts_api_v1_simulation_runs__run_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationArtifactsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Artifacts",
        "tags": ["simulation-workbench"]
      }
    },
    "/api/v1/simulation/runs/{run_id}/attempts": {
      "get": {
        "operationId": "run_attempts_api_v1_simulation_runs__run_id__attempts_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Attempts",
        "tags": ["simulation-workbench"]
      }
    },
    "/api/v1/simulation/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_api_v1_simulation_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": ["simulation-workbench"]
      }
    },
    "/api/v1/simulation/runs/{run_id}/clone": {
      "post": {
        "operationId": "clone_run_api_v1_simulation_runs__run_id__clone_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReproductionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clone Run",
        "tags": ["simulation-workbench"]
      }
    },
    "/api/v1/simulation/runs/{run_id}/events": {
      "get": {
        "operationId": "run_events_api_v1_simulation_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Events",
        "tags": ["simulation-workbench"]
      }
    },
...
    "/api/v1/simulation/runs/{run_id}/reproduce": {
      "post": {
        "operationId": "reproduce_run_api_v1_simulation_runs__run_id__reproduce_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReproductionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reproduce Run",
        "tags": ["simulation-workbench"]
      }
    },
    "/api/v1/simulation/runs/{run_id}/retry": {
      "post": {
        "operationId": "retry_run_api_v1_simulation_runs__run_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Run",
        "tags": ["simulation-workbench"]
      }
    },
...
    "/api/v1/simulation/scenarios/{scenario_id}": {
      "get": {
        "operationId": "scenario_api_v1_simulation_scenarios__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDefinitionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario",
        "tags": ["simulation-workbench"]
      }
    },
...
    "/api/v1/simulation/validate": {
      "post": {
        "operationId": "validate_api_v1_simulation_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate",
        "tags": ["simulation-workbench"]
      }
    },
...
  }
}